// src/app/jama/model/requestTestCycle.ts
/**
 * Jama REST API
 * This is the documentation for the Jama REST API.
 *
 * OpenAPI spec version: latest
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { TestRunGenerationConfig } from './testRunGenerationConfig';

export interface RequestTestCycle {
	/**
	 * A map of field names to field values e.g. {\"name\":\"Sample Item\", \"startDate\": \"2015-01-01\", \"endDate\": \"2015-01-15\"} Note: RequestTestCycle requires name, startDate, and endDate. startDate and endDate are formatted with 'yyyy-mm-dd'
	 */
	fields: { [key: string]: any; };

	/**
	 * Settings for how test runs will be generated in this test cycle
	 */
	testRunGenerationConfig: TestRunGenerationConfig;

}